requirements.txt
pandas>=2.0.0
psycopg2-binary>=2.9.0
openai>=1.0.0
datasets>=2.0.0
jsonschema
python-dotenv
sentence-transformers>=2.2.0
torch>=2.0.0
vllm>=0.6.0